requirements.txt
asn1crypto==0.24.0
azure-common==1.1.23
azure-functions==1.0.0b5
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
idna==2.8
pycparser==2.19
python-dateutil==2.8.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3
